Split multi-class files
