+++ Sources/Mailozaurr/Attachment.cs
@@ -1,0 +1,10 @@
+namespace Mailozaurr;
+
+/// <summary>
+/// Represents a file attachment from Microsoft Graph.
+/// </summary>
+public class Attachment {
+    public string Name { get; set; }
+    public string ContentBytes { get; set; }
+    // Add more properties as needed
+}
