+++ Sources/Mailozaurr/Helpers/DataProtectionScope.cs
@@ -1,0 +1,8 @@
+#if !UNIX
+namespace Mailozaurr;
+
+internal enum DataProtectionScope {
+    CurrentUser = 0x00,
+    LocalMachine = 0x01
+}
+#endif
