+++ Sources/Mailozaurr/Helpers/SecureStringHelper.cs
@@ -336,229 +336,4 @@ namespace Mailozaurr {
 #nullable restore
     }
 
-    /// <summary>
-    /// Helper class to return encryption results, and the IV used to
-    /// do the encryption.
-    /// </summary>
-    internal class EncryptionResult {
-        internal EncryptionResult(string encrypted, string IV) {
-            EncryptedData = encrypted;
-            this.IV = IV;
-        }
-
-        /// <summary>
-        /// Gets the encrypted data.
-        /// </summary>
-        internal string EncryptedData { get; }
-
-        /// <summary>
-        /// Gets the IV used to encrypt the data.
-        /// </summary>
-        internal string IV { get; }
-    }
-
-#if !UNIX
-
-    // The DPAPIs implemented in this section are temporary workaround.
-    // CoreCLR team will bring 'ProtectedData' type to Project K eventually.
-
-    #region DPAPI
-
-    internal enum DataProtectionScope {
-        CurrentUser = 0x00,
-        LocalMachine = 0x01
-    }
-
-    internal static class ProtectedData {
-        /// <summary>
-        /// Protect.
-        /// </summary>
-        public static byte[] Protect(byte[] userData, byte[] optionalEntropy, DataProtectionScope scope) {
-            //ArgumentNullException.ThrowIfNull(userData);
-            if (userData.Length == 0) {
-                throw new ArgumentException("Cryptography_DpApi_InvalidDataToProtect", nameof(userData));
-            }
-
-            GCHandle pbDataIn = new GCHandle();
-            GCHandle pOptionalEntropy = new GCHandle();
-            CAPI.CRYPTOAPI_BLOB blob = new CAPI.CRYPTOAPI_BLOB();
-
-            try {
-                pbDataIn = GCHandle.Alloc(userData, GCHandleType.Pinned);
-                CAPI.CRYPTOAPI_BLOB dataIn = new CAPI.CRYPTOAPI_BLOB();
-                dataIn.cbData = (uint)userData.Length;
-                dataIn.pbData = pbDataIn.AddrOfPinnedObject();
-                CAPI.CRYPTOAPI_BLOB entropy = new CAPI.CRYPTOAPI_BLOB();
-                if (optionalEntropy != null) {
-                    pOptionalEntropy = GCHandle.Alloc(optionalEntropy, GCHandleType.Pinned);
-                    entropy.cbData = (uint)optionalEntropy.Length;
-                    entropy.pbData = pOptionalEntropy.AddrOfPinnedObject();
-                }
-
-                uint dwFlags = CAPI.CRYPTPROTECT_UI_FORBIDDEN;
-                if (scope == DataProtectionScope.LocalMachine)
-                    dwFlags |= CAPI.CRYPTPROTECT_LOCAL_MACHINE;
-                unsafe {
-                    if (!CAPI.CryptProtectData(
-                        pDataIn: new IntPtr(&dataIn),
-                        szDataDescr: string.Empty,
-                        pOptionalEntropy: new IntPtr(&entropy),
-                        pvReserved: IntPtr.Zero,
-                        pPromptStruct: IntPtr.Zero,
-                        dwFlags: dwFlags,
-                        pDataBlob: new IntPtr(&blob))) {
-                        int lastWin32Error = Marshal.GetLastWin32Error();
-
-                        // One of the most common reasons that DPAPI operations fail is that the user
-                        // profile is not loaded (for instance in the case of impersonation or running in a
-                        // service.  In those cases, throw an exception that provides more specific details
-                        // about what happened.
-                        if (CAPI.ErrorMayBeCausedByUnloadedProfile(lastWin32Error)) {
-                            throw new CryptographicException("Cryptography_DpApi_ProfileMayNotBeLoaded");
-                        } else {
-                            throw new CryptographicException(lastWin32Error);
-                        }
-                    }
-                }
-
-                // In some cases, the API would fail due to OOM but simply return a null pointer.
-                if (blob.pbData == IntPtr.Zero) {
-                    throw new OutOfMemoryException();
-                }
-
-                byte[] encryptedData = new byte[(int)blob.cbData];
-                Marshal.Copy(blob.pbData, encryptedData, 0, encryptedData.Length);
-
-                return encryptedData;
-            } finally {
-                if (pbDataIn.IsAllocated) {
-                    pbDataIn.Free();
-                }
-                if (pOptionalEntropy.IsAllocated) {
-                    pOptionalEntropy.Free();
-                }
-                if (blob.pbData != IntPtr.Zero) {
-                    CAPI.ZeroMemory(blob.pbData, blob.cbData);
-                    CAPI.LocalFree(blob.pbData);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Unprotect.
-        /// </summary>
-        public static byte[] Unprotect(byte[] encryptedData, byte[] optionalEntropy, DataProtectionScope scope) {
-            //ArgumentNullException.ThrowIfNull(encryptedData);
-            if (encryptedData.Length == 0) {
-                throw new ArgumentException("Cryptography_DpApi_InvalidDataToUnprotect", nameof(encryptedData));
-            }
-
-            GCHandle pbDataIn = new GCHandle();
-            GCHandle pOptionalEntropy = new GCHandle();
-            CAPI.CRYPTOAPI_BLOB userData = new CAPI.CRYPTOAPI_BLOB();
-
-            try {
-                pbDataIn = GCHandle.Alloc(encryptedData, GCHandleType.Pinned);
-                CAPI.CRYPTOAPI_BLOB dataIn = new CAPI.CRYPTOAPI_BLOB();
-                dataIn.cbData = (uint)encryptedData.Length;
-                dataIn.pbData = pbDataIn.AddrOfPinnedObject();
-                CAPI.CRYPTOAPI_BLOB entropy = new CAPI.CRYPTOAPI_BLOB();
-                if (optionalEntropy != null) {
-                    pOptionalEntropy = GCHandle.Alloc(optionalEntropy, GCHandleType.Pinned);
-                    entropy.cbData = (uint)optionalEntropy.Length;
-                    entropy.pbData = pOptionalEntropy.AddrOfPinnedObject();
-                }
-
-                uint dwFlags = CAPI.CRYPTPROTECT_UI_FORBIDDEN;
-                if (scope == DataProtectionScope.LocalMachine) {
-                    dwFlags |= CAPI.CRYPTPROTECT_LOCAL_MACHINE;
-                }
-
-                unsafe {
-                    if (!CAPI.CryptUnprotectData(
-                        pDataIn: new IntPtr(&dataIn),
-                        ppszDataDescr: IntPtr.Zero,
-                        pOptionalEntropy: new IntPtr(&entropy),
-                        pvReserved: IntPtr.Zero,
-                        pPromptStruct: IntPtr.Zero,
-                        dwFlags: dwFlags,
-                        pDataBlob: new IntPtr(&userData))) {
-                        throw new CryptographicException(Marshal.GetLastWin32Error());
-                    }
-                }
-
-                // In some cases, the API would fail due to OOM but simply return a null pointer.
-                if (userData.pbData == IntPtr.Zero) {
-                    throw new OutOfMemoryException();
-                }
-
-                byte[] data = new byte[(int)userData.cbData];
-                Marshal.Copy(userData.pbData, data, 0, data.Length);
-
-                return data;
-            } finally {
-                if (pbDataIn.IsAllocated) {
-                    pbDataIn.Free();
-                }
-                if (pOptionalEntropy.IsAllocated) {
-                    pOptionalEntropy.Free();
-                }
-                if (userData.pbData != IntPtr.Zero) {
-                    CAPI.ZeroMemory(userData.pbData, userData.cbData);
-                    CAPI.LocalFree(userData.pbData);
-                }
-            }
-        }
-    }
-    internal static class CAPI {
-        internal const uint CRYPTPROTECT_UI_FORBIDDEN = 0x1;
-        internal const uint CRYPTPROTECT_LOCAL_MACHINE = 0x4;
-
-        internal const int E_FILENOTFOUND = unchecked((int)0x80070002); // File not found
-        internal const int ERROR_FILE_NOT_FOUND = 2;                    // File not found
-
-        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-        internal struct CRYPTOAPI_BLOB {
-            internal uint cbData;
-            internal IntPtr pbData;
-        }
-
-        internal static bool ErrorMayBeCausedByUnloadedProfile(int errorCode) {
-            // CAPI returns a file not found error if the user profile is not yet loaded
-            return errorCode == E_FILENOTFOUND ||
-                   errorCode == ERROR_FILE_NOT_FOUND;
-        }
-
-        [DllImport("CRYPT32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        internal static extern bool CryptProtectData(
-                [In] IntPtr pDataIn,
-                [In] string szDataDescr,
-                [In] IntPtr pOptionalEntropy,
-                [In] IntPtr pvReserved,
-                [In] IntPtr pPromptStruct,
-                [In] uint dwFlags,
-                [In, Out] IntPtr pDataBlob);
-
-        [DllImport("CRYPT32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        internal static extern bool CryptUnprotectData(
-                [In] IntPtr pDataIn,
-                [In] IntPtr ppszDataDescr,
-                [In] IntPtr pOptionalEntropy,
-                [In] IntPtr pvReserved,
-                [In] IntPtr pPromptStruct,
-                [In] uint dwFlags,
-                [In, Out] IntPtr pDataBlob);
-
-        [DllImport("ntdll.dll", EntryPoint = "RtlZeroMemory", SetLastError = true)]
-        internal static extern void ZeroMemory(IntPtr handle, uint length);
-
-        [DllImport("api-ms-win-core-misc-l1-1-0.dll", SetLastError = true)]
-        internal static extern IntPtr LocalFree(IntPtr handle);
-    }
-
-    #endregion DPAPI
-
-#endif
 }
