+++ Sources/Mailozaurr/Logging/InternalLogger.cs
@@ -209,83 +209,3 @@ public class InternalLogger {
 /// Represents the arguments for a log event.
 /// </summary>
 /// <seealso cref="System.EventArgs" />
-public class LogEventArgs : EventArgs {
-    /// <summary>
-    /// Progress percentage
-    /// </summary>
-    public int? ProgressPercentage { get; set; }
-
-    /// <summary>
-    /// Progress total steps
-    /// </summary>
-    public int? ProgressTotalSteps { get; set; }
-
-    /// <summary>
-    /// Progress current steps
-    /// </summary>
-    public int? ProgressCurrentSteps { get; set; }
-
-    /// <summary>
-    /// Progress current operation
-    /// </summary>
-    public string ProgressCurrentOperation { get; set; }
-
-    /// <summary>
-    /// Progress activity
-    /// </summary>
-    public string ProgressActivity { get; set; }
-
-    /// <summary>
-    /// Message to be written including arguments substitution
-    /// </summary>
-    public string FullMessage { get; set; }
-
-    /// <summary>
-    /// Message to be written
-    /// </summary>
-    public string Message { get; set; }
-
-    /// <summary>
-    /// Gets or sets the arguments.
-    /// </summary>
-    /// <value>
-    /// The arguments.
-    /// </value>
-    public object[] Args { get; set; }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LogEventArgs"/> class.
-    /// </summary>
-    /// <param name="message">The message.</param>
-    /// <param name="args">The arguments.</param>
-    public LogEventArgs(string message, object[] args) {
-        Message = message;
-        Args = args;
-        FullMessage = string.Format(message, args);
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LogEventArgs"/> class.
-    /// </summary>
-    /// <param name="message">The message.</param>
-    public LogEventArgs(string message) {
-        Message = message;
-        FullMessage = message;
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LogEventArgs"/> class.
-    /// </summary>
-    /// <param name="activity">The activity.</param>
-    /// <param name="currentOperation">The current operation.</param>
-    /// <param name="currentSteps">The current steps.</param>
-    /// <param name="totalSteps">The total steps.</param>
-    /// <param name="percentage">The percentage.</param>
-    public LogEventArgs(string activity, string currentOperation, int? currentSteps, int? totalSteps, int? percentage) {
-        ProgressActivity = activity;
-        ProgressCurrentOperation = currentOperation;
-        ProgressCurrentSteps = currentSteps;
-        ProgressTotalSteps = totalSteps;
-        ProgressPercentage = percentage;
-    }
-}
