+++ Sources/Mailozaurr/MicrosoftGraph/GraphAuthorization.cs
@@ -3,14 +3,6 @@ using System.Text.Json.Serialization;
 namespace Mailozaurr;
 
 /// <summary>
-/// Result returned when creating an upload session for large attachments.
-/// </summary>
-public class GraphUploadSessionResult {
-    [JsonPropertyName("uploadUrl")]
-    public string UploadUrl { get; set; }
-}
-
-/// <summary>
 /// Authorization information returned by Microsoft Graph OAuth flows.
 /// </summary>
 public class GraphAuthorization {
