+++ Sources/Mailozaurr/MicrosoftGraph/GraphEmail.cs
@@ -1,0 +1,9 @@
+namespace Mailozaurr;
+
+/// <summary>
+/// Simple email address container.
+/// </summary>
+public class GraphEmail {
+    [JsonPropertyName("address")]
+    public string Address { get; set; }
+}
