+++ Sources/Mailozaurr/MicrosoftGraph/GraphEmailAddress.cs
@@ -1,0 +1,9 @@
+namespace Mailozaurr;
+
+/// <summary>
+/// Represents an email address object for Graph API payloads.
+/// </summary>
+public class GraphEmailAddress {
+    [JsonPropertyName("emailAddress")]
+    public GraphEmail Email { get; set; }
+}
